feat(handler): add Multiply handler
